get-started-checkpoints/code-checkpoints/00-common/requirements.txt
tensorflow>=2.5,<2.6
ruamel.yaml
dvc[all]>=2.1
